L2 devchain fix (#11200)
* L2 devchain fix

* fixes

* artifact fixes

+++ packages/protocol/scripts/foundry/create_and_migrate_anvil_devchain.sh
@@ -64,6 +64,11 @@ forge script \
 ELAPSED_TIME=$(($SECONDS - $START_TIME))
 echo "Total elapsed time: $ELAPSED_TIME seconds"
 
-# Rename devchain artifact and remove unused directory
-mv $ANVIL_FOLDER/state.json $TMP_FOLDER/$L1_DEVCHAIN_FILE_NAME
-rm -rf $ANVIL_FOLDER
+if [[ "${KEEP_DEVCHAIN_FOLDER:-}" == "true" ]]; then
+    cp $ANVIL_FOLDER/state.json $TMP_FOLDER/$L1_DEVCHAIN_FILE_NAME
+    echo "Keeping devchain folder as per flag."
+else
+    # Rename devchain artifact and remove unused directory
+    mv $ANVIL_FOLDER/state.json $TMP_FOLDER/$L1_DEVCHAIN_FILE_NAME
+    rm -rf $ANVIL_FOLDER
+fi
